Abstract enum base into util
